Handle ALLOWED_VALUES_PROPERTY in propertyValue
